Various fixes  - Updated README  - Updated folder structure  - Introduced .editorconfig  - Fixed a few frontend issues
