+++ README.md
@@ -15,53 +15,49 @@ This is an opensource blockchain prototype project that aims to solve some block
 
 ## Folder Structure
 
--baby-blockchain
+baby-blockchain
+│
+├── backend
+│   ├── AddNet              - TODO: add decentralization and fault tolerance
+│   ├── AddPersistence      - TODO: add data persistance
+│   ├── AddPoS              - TODO: Proof of Stake concensus algorithm
+│   └── Simple              - First and Simple implementation (No data persistence: Data disapears after reload) built in Go
+└── frontend                - Frontend built in React (js)
 
-  - baby-blockchain
-      - AddNet : add decentralization and fault tolerance 
-      - AddPersistence : data persistence 
-      - AddPoS : Proof of Stake concensus algorithm
-      - Simple : First and Simple implementation (No data persistence: Data disapears after reload)
-
-  - frontend : Reactjs frontend App
 
 ## Prerequisites
 
-- go installed
-- nodejs installed
+ - go installed
+ - nodejs installed
  
 ## Run it
 
-- git clone https://github.com/ndaysinaiK/baby-blockchain
-- cd baby-blockchain/frontend
-- npm i
-- cd baby-blockchain/baby-blockchain/Simple : for the first implementation
-- go build
-- run 
-- run ./blockchain : this will run the app on port 3002 ( may shut down any app running on that port)
-- open the frontend folder, search for https://bch-simple.herokuapp.com in Utils/Apis.js and replace it with http://localhost:3002
-- go to http://localhost:3002/api to see your api running.
-- cd baby-blockchain/frontend 
-- npm start : this will open the reactjs app in your local browser. Test and improve it.
+### Backend
 
+```sh
+cd backend/Simple
+cp dotenv.tpl .env                  # create a .env file from template
+go build                            # build application
+./blockchain                        # run the application on the port specified in .env (defaults to 3002 - may shut down any app running on that port)
+```
+
+Go to `http://localhost:3002/api` to see your api running.
+
+### Frontend
+
+```sh
+cd frontend
+npm install
+npm start                           # this will open the reactjs app in your default browser (localhost:3000 by default).
+```
+
+Go to http://localhost:3000 to see your frontend running.
+Note: Backend url is specified in `Utils/Apis.js`
 
 ### This project is an improvement of this article https://medium.com/swlh/is-it-hard-to-build-a-blockchain-from-scratch-2662e9b873b7
 
 You can read it for basic understanding.
 
-
 ## Any contribution will be appreciated
 
 Fork the repo, pull requests to contribute to this project.
-
-
-
-
-
-
-
-
-
-
-
-
