--- baby-blockchain/AddPoS/README.md
@@ -1,1 +1,0 @@
-# add proof of state or other concensus algorithms
