+++ backend/AddPoS/README.md
@@ -1,0 +1,1 @@
+# add proof of state or other concensus algorithms
