+++ backend/Simple/dotenv.tpl
@@ -1,0 +1,3 @@
+# Backend http port
+PORT=3002
+
