+++ backend/Simple/go.mod
@@ -1,0 +1,14 @@
+module blockchain
+
+go 1.17
+
+require (
+	github.com/davecgh/go-spew v1.1.1
+	github.com/gorilla/mux v1.8.0
+	github.com/joho/godotenv v1.4.0
+)
+
+require (
+	github.com/felixge/httpsnoop v1.0.1 // indirect
+	github.com/gorilla/handlers v1.5.1 // indirect
+)
